+++ src/test/resources/db/component/component_link_create_mysql.sql
@@ -11,8 +11,8 @@ CREATE TABLE IF NOT EXISTS component_links
     component_id           bigint       NOT NULL,
     user_id                bigint   DEFAULT NULL COMMENT 'the owner / creator of the component_link',
     order_nbr              bigint       NOT NULL,
-    component_link_type_id bigint   NOT NULL DEFAULT 1,
-    position_type_id       bigint   NOT NULL DEFAULT 2 COMMENT 'the position of the component e.g. right or below',
+    component_link_type_id smallint NOT NULL DEFAULT 1,
+    position_type_id       smallint NOT NULL DEFAULT 2 COMMENT 'the position of the component e.g. right or below',
     excluded               smallint DEFAULT NULL COMMENT 'true if a user,but not all,have removed it',
     share_type_id          smallint DEFAULT NULL COMMENT 'to restrict the access',
     protect_id             smallint DEFAULT NULL COMMENT 'to protect against unwanted changes'
@@ -36,8 +36,8 @@ CREATE TABLE IF NOT EXISTS user_component_links
     component_link_id      bigint       NOT NULL COMMENT 'with the user_id the internal unique primary index',
     user_id                bigint       NOT NULL COMMENT 'the changer of the component_link',
     order_nbr              bigint   DEFAULT NULL,
-    component_link_type_id bigint   DEFAULT NULL,
-    position_type_id       bigint   DEFAULT NULL COMMENT 'the position of the component e.g. right or below',
+    component_link_type_id smallint DEFAULT NULL,
+    position_type_id       smallint DEFAULT NULL COMMENT 'the position of the component e.g. right or below',
     excluded               smallint DEFAULT NULL COMMENT 'true if a user,but not all,have removed it',
     share_type_id          smallint DEFAULT NULL COMMENT 'to restrict the access',
     protect_id             smallint DEFAULT NULL COMMENT 'to protect against unwanted changes'
