+++ src/test/resources/db/log/change_big_value_create.sql
@@ -11,7 +11,7 @@ CREATE TABLE IF NOT EXISTS change_big_values
     user_id          bigint    NOT NULL,
     change_action_id smallint  NOT NULL,
     group_id         text      NOT NULL,
-    change_field_id  bigint    NOT NULL,
+    change_field_id  smallint  NOT NULL,
     old_value        double precision DEFAULT NULL,
     new_value        double precision DEFAULT NULL
 );
