+++ src/test/resources/db/sys_log/system_time_create_mysql.sql
@@ -6,11 +6,11 @@
 
 CREATE TABLE IF NOT EXISTS system_times
 (
-    system_time_id bigint NOT NULL COMMENT 'the internal unique primary index',
-    start_time timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP COMMENT 'start time of the monitoring period',
-    end_time timestamp DEFAULT NULL COMMENT 'end time of the monitoring period',
-    system_time_type_id bigint NOT NULL COMMENT 'the area of the execution time e.g. db write',
-    milliseconds bigint NOT NULL COMMENT 'the execution time in milliseconds'
+    system_time_id      bigint        NOT NULL COMMENT 'the internal unique primary index',
+    start_time          timestamp     NOT NULL DEFAULT CURRENT_TIMESTAMP COMMENT 'start time of the monitoring period',
+    end_time            timestamp DEFAULT NULL COMMENT 'end time of the monitoring period',
+    system_time_type_id smallint      NOT NULL COMMENT 'the area of the execution time e.g. db write',
+    milliseconds        bigint        NOT NULL COMMENT 'the execution time in milliseconds'
 )
     ENGINE = InnoDB
     DEFAULT CHARSET = utf8
